+++ api/src/main/java/io/github/ysdaeth/jmodularcrypt/api/Encryptor.java
@@ -1,5 +1,6 @@
 package io.github.ysdaeth.jmodularcrypt.api;
 
+import java.security.Key;
 import java.security.KeyException;
 
 /**
@@ -13,22 +14,24 @@ public interface Encryptor {
     /**
      * Encrypt given character sequence and return its representation
      * in Modular Crypt Format
-     * @param secret credentials to be encrypted
+     * @param secret secret to be encrypted
+     * @param encryptionKey key to encrypt the secret
      * @return Modular Crypt Format string value
      * @throws KeyException when key is invalid, missing or not recognized
      */
-    String encrypt(byte[] secret) throws KeyException;
+    String encrypt(byte[] secret, Key encryptionKey) throws KeyException;
 
     /**
      * Decrypt given Modular Crypt Format string representation
      * and return decrypted value
      * @param encrypted Modular Crypt Format string produced by the algorithm.
-     * @return decrypted credentials
+     * @param decryptionKey key to decrypt encrypted data
+     * @return decrypted secret
      * @throws KeyException when key is invalid, missing or not recognized
      */
-    byte[] decrypt(String encrypted) throws KeyException;
+    byte[] decrypt(String encrypted, Key decryptionKey) throws KeyException;
 
-    /**\
+    /**
      * Returns the cryptographic algorithm identifier.
      * Identifier should be first section of Modular Crypt Format
      * @return algorithm identifier