+++ implementation/src/main/java/io/github/ysdaeth/jmodularcrypt/impl/encryptor/IncorrectAlgorithmException.java
@@ -1,0 +1,12 @@
+package io.github.ysdaeth.jmodularcrypt.impl.encryptor;
+
+public class IncorrectAlgorithmException extends RuntimeException {
+    public IncorrectAlgorithmException(String message) {
+        super(message);
+    }
+
+    public IncorrectAlgorithmException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}